requirements.txt
Babel==2.3.4
Django==1.9
Faker==0.7.7
Pillow==3.4.2
SQLAlchemy==0.9.9
Unidecode==0.04.19
argparse==1.2.1
boto==2.45.0
chardet==2.3.0
cloudinary==1.1.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.1
django-autoslug==1.9.3
django-bootstrap3==7.0.0
django-compressor==2.1
django-extra-views==0.6.4
django-haystack==2.5.1
django-jsonfield==0.9.13
django-localflavor==1.3
django-material==0.10.1
django-oscar==1.3
django-oscar-api==1.0.10-1
django-oscar-paymentexpress==0.1.1
django-oscar-paypal==0.9.7
django-rest-auth==0.9.0
django-storages==1.5.1
django-tables2==1.0.7
django-tastypie==0.13.3
django-toolbelt==0.0.1
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
factory-boy==2.8.1
funcsigs==1.0.2
gunicorn==19.3.0
ipaddress==1.0.17
lxml==3.6.1
mock==2.0.0
multidict==1.2.1
parse==1.6.6
pbr==1.10.0
phonenumbers==7.7.5
progressbar2==3.9.4
psycopg2==2.6.1
purl==1.3
pyasn1==0.1.9
pycountry==16.11.27.1
pycrypto==2.6.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
python-utils==2.0.0
pytz==2015.7
rcssmin==1.0.6
requests==2.6.0
rjsmin==1.0.12
rsa==3.3
simplejson==3.8.2
six==1.10.0
sorl-thumbnail==12.4a1
static3==0.5.1
tzlocal==1.2.2
wsgiref==0.1.2
